ixpe generation iteration with ephem generalization

+++ migrations/versions/_2025_05_14_1521-a6590318e932_create_ixpe.py
@@ -8,7 +8,6 @@ Create Date: 2025-05-14 15:21:03.729061
 
 from __future__ import annotations
 
-import uuid
 from typing import Sequence, Union
 from uuid import uuid4
 
@@ -18,30 +17,14 @@ from alembic import op
 from sqlalchemy import orm, select
 
 from across_server.core.enums import EphemerisType, InstrumentFOV, InstrumentType
-from migrations.db_util import ACROSSFootprintPoint, create_geography
-from migrations.versions.model_snapshots.models_2025_05_15 import (
-    Filter,
-    Footprint,
-    Group,
-    GroupRole,
-    Instrument,
-    Observatory,
-    ObservatoryEphemerisType,
-    Permission,
-    Telescope,
-    TLEParameters,
-)
+from migrations.db_util import ACROSSFootprintPoint, arcmin_to_deg, create_geography
+from migrations.versions.model_snapshots import models_2025_05_15 as model_snapshots
 
 # revision identifiers, used by Alembic.
 revision: str = "a6590318e932"
 down_revision: Union[str, None] = "043885e1cd78"
 branch_labels: Union[str, Sequence[str], None] = None
 depends_on: Union[str, Sequence[str], None] = None
-
-
-def arcmin_to_deg(arcmin: float) -> float:
-    """Convert arcminute to degrees."""
-    return arcmin / 60.0
 
 
 # 12 arcmin x 12 arcmin square
@@ -65,7 +48,7 @@ IXPE_WAVELENGTH_BANDPASS = convert_to_wave(IXPE_ENERGY_BANDPASS)
 OBSERVATORY = {
     "name": "Imaging X-Ray Polarimetry Explorer",
     "short_name": "IXPE",
-    "observatory_type": "SPACE_BASED",
+    "type": "SPACE_BASED",
     "reference_url": "https://ixpe.msfc.nasa.gov/index.html",
     "is_operational": True,
     "telescopes": [
@@ -96,6 +79,16 @@ OBSERVATORY = {
             ],
         }
     ],
+    "ephemeris_types": [
+        {
+            "ephemeris_type": EphemerisType.TLE,
+            "priority": 1,
+            "parameters": {
+                "norad_id": 49954,
+                "norad_satellite_name": "IXPE",
+            },
+        }
+    ],
 }
 
 
@@ -103,84 +96,66 @@ def upgrade() -> None:
     bind = op.get_bind()
     session = orm.Session(bind=bind, expire_on_commit=False)
 
-    # create group based off the observatory
-    group = Group(
+    group_permissions = (
+        session.query(model_snapshots.Permission)
+        .filter(model_snapshots.Permission.name.contains("group"))
+        .all()
+    )
+    # create the group
+    group = model_snapshots.Group(
         id=uuid4(), name=OBSERVATORY["name"], short_name=OBSERVATORY["short_name"]
     )
     session.add(group)
 
     # create group admin
-    permissions = (
-        session.query(Permission)
-        .filter(
-            Permission.name.in_(
-                [
-                    "group:user:write",
-                    "group:role:write",
-                    "group:write",
-                    "group:read",
-                    "group:observatory:write",
-                    "group:telescope:write",
-                    "group:schedule:write",
-                ]
-            )
-        )
-        .all()
-    )
-    group_admin = GroupRole(
-        id=uuid.UUID("ae1af917-093b-4b2b-a11d-cb1097470e90"),
-        name="IXPE Group Admin",
-        permissions=permissions,
+    group_admin = model_snapshots.GroupRole(
+        id=uuid4(),
+        name=f"{group.short_name} Group Admin",
+        permissions=group_permissions,
         group=group,
     )
     session.add(group_admin)
 
-    # # create observatory
-    observatory_insert = Observatory(
+    # observatory preparation
+    telescopes = OBSERVATORY.pop("telescopes")
+    ephemeris_types = OBSERVATORY.pop("ephemeris_types")
+
+    # create the observatory
+    observatory_insert = model_snapshots.Observatory(
         id=uuid4(),
-        name=OBSERVATORY["name"],
-        short_name=OBSERVATORY["short_name"],
-        type=OBSERVATORY["observatory_type"],
-        reference_url=OBSERVATORY["reference_url"],
-        is_operational=OBSERVATORY["is_operational"],
         group=group,
+        **OBSERVATORY,
     )
     session.add(observatory_insert)
     observatory_id = observatory_insert.id
 
-    # get the telescopes
-    telescopes = OBSERVATORY["telescopes"]
+    # iterate over the telescopes
     for telescope in telescopes:  #  type:ignore
+        instruments = telescope.pop("instruments")
+
         # create the telescope record
-        telescope_insert = Telescope(
+        telescope_insert = model_snapshots.Telescope(
             id=uuid4(),
-            name=telescope["name"],
-            short_name=telescope["short_name"],
-            reference_url=telescope["reference_url"],
-            is_operational=telescope["is_operational"],
             observatory_id=observatory_id,
+            **telescope,
         )
         session.add(telescope_insert)
         telescope_id = telescope_insert.id
 
         # get the instruments
-        instruments = telescope["instruments"]  #  type:ignore
         for instrument in instruments:
+            footprints = instrument.pop("footprint")
+            filters = instrument.pop("filters")
+
             # create the instrument record
-            instrument_insert = Instrument(
+            instrument_insert = model_snapshots.Instrument(
                 id=uuid4(),
-                name=instrument["name"],
-                short_name=instrument["short_name"],
-                type=instrument["type"],
-                reference_url=instrument["reference_url"],
-                is_operational=instrument["is_operational"],
-                field_of_view=instrument["field_of_view"],
                 telescope_id=telescope_id,
+                **instrument,
             )
             session.add(instrument_insert)
             instrument_id = instrument_insert.id
 
-            footprints = instrument["footprint"]
             for footprint in footprints:
                 # convert the input footprint to the string polygon
                 vertices = []
@@ -190,31 +165,51 @@ def upgrade() -> None:
                 polygon = create_geography(polygon=vertices)
 
                 # create the footprint model record
-                footprint_insert = Footprint(
-                    polygon=polygon, instrument_id=instrument_id
+                footprint_insert = model_snapshots.Footprint(
+                    instrument_id=instrument_id, polygon=polygon
                 )
                 session.add(footprint_insert)
 
-            filters = instrument["filters"]
             for filter in filters:
-                filter_insert = Filter(**filter)
-                filter_insert.instrument_id = instrument_id
+                # create the filters
+                filter_insert = model_snapshots.Filter(
+                    instrument_id=instrument_id, **filter
+                )
                 session.add(filter_insert)
 
-    # Add TLEParameters for IXPE
-    ephemeris_type = ObservatoryEphemerisType(
-        observatory_id=observatory_id,
-        ephemeris_type=EphemerisType.TLE,
-        priority=1,
-    )
-    session.add(ephemeris_type)
+    for ephemeris_type in ephemeris_types:  # type:ignore
+        # create the ephemeris type record
+        parameters = ephemeris_type.pop("parameters")
+        ephemeris_type_insert = model_snapshots.ObservatoryEphemerisType(
+            id=uuid4(),
+            observatory_id=observatory_id,
+            **ephemeris_type,
+        )
+        session.add(ephemeris_type_insert)
 
-    tle_parameters = TLEParameters(
-        observatory_id=observatory_id,
-        norad_id=49954,
-        norad_satellite_name="IXPE",
-    )
-    session.add(tle_parameters)
+        if ephemeris_type_insert.ephemeris_type == EphemerisType.TLE:
+            tle_parameters = model_snapshots.TLEParameters(
+                id=uuid4(), observatory_id=observatory_id, **parameters
+            )
+            session.add(tle_parameters)
+
+        if ephemeris_type_insert.ephemeris_type == EphemerisType.JPL:
+            jpl_parameters = model_snapshots.JPLEphemerisParameters(
+                id=uuid4(), observatory_id=observatory_id, **parameters
+            )
+            session.add(jpl_parameters)
+
+        if ephemeris_type_insert.ephemeris_type == EphemerisType.SPICE:
+            spice_parameters = model_snapshots.SpiceKernelParameters(
+                id=uuid4(), observatory_id=observatory_id, **parameters
+            )
+            session.add(spice_parameters)
+
+        if ephemeris_type_insert.ephemeris_type == EphemerisType.GROUND:
+            ground_parameters = model_snapshots.EarthLocationParameters(
+                id=uuid4(), observatory_id=observatory_id, **parameters
+            )
+            session.add(ground_parameters)
 
     session.commit()
 
@@ -223,17 +218,46 @@ def downgrade() -> None:
     bind = op.get_bind()
     session = orm.Session(bind=bind)
 
-    group = session.scalar(select(Group).filter_by(name=OBSERVATORY["name"]))
     observatory = session.scalar(
-        select(Observatory).filter_by(name=OBSERVATORY["name"])
-    )
-    tle_parameters = session.scalar(
-        select(TLEParameters).filter_by(observatory_id=observatory.id)  # type:ignore
+        select(model_snapshots.Observatory).filter_by(name=OBSERVATORY["name"])  # type:ignore
     )
 
-    session.delete(group)
+    group = session.scalar(
+        select(model_snapshots.Group).filter_by(name=observatory.name)  # type:ignore
+    )
+
     session.delete(observatory)
-    session.delete(tle_parameters)
+    session.delete(group)
+
+    tle_parameters = session.scalar(
+        select(model_snapshots.TLEParameters).filter_by(observatory_id=observatory.id)  # type:ignore
+    )
+    if tle_parameters:
+        session.delete(tle_parameters)
+
+    jpl_parameters = session.scalar(
+        select(model_snapshots.JPLEphemerisParameters).filter_by(
+            observatory_id=observatory.id  # type:ignore
+        )
+    )
+    if jpl_parameters:
+        session.delete(jpl_parameters)
+
+    spice_parameters = session.scalar(
+        select(model_snapshots.SpiceKernelParameters).filter_by(
+            observatory_id=observatory.id  # type:ignore
+        )
+    )
+    if spice_parameters:
+        session.delete(spice_parameters)
+
+    ground_parameters = session.scalar(
+        select(model_snapshots.EarthLocationParameters).filter_by(
+            observatory_id=observatory.id  # type:ignore
+        )
+    )
+    if ground_parameters:
+        session.delete(ground_parameters)
 
     session.commit()
     # fin
